Move ORM machinery from module level to KB session level
The previous ORM style used by the knowledge base was "global", in the
sense that several key elements were shared at the Python module level.
As a consequence, even if the concept of "KB session" was provided, it
was actually impossible to manage two independent connections to two
distinct knowledge bases at the same time: the system did not support
handling different table prefixes and DB schemas.

This changeset binds the whole ORM machinery to the KB session.  The
main result is that KBClass, KBObject and Attribute classes are built
when a Session is created, together with their SQLAlchemy mappers and
events.

The tests and Django view files have been updated, following the slight
API changes introduced by this enhancement.